Create HW
 I dont know to make the code print the current small/ large integer.

Each iteration will  create a random integer.  Examine each integer and, if it is either larger than the current large or smaller than the current small, record the new large or small integer.  These will almost certainly change as the loop executes.
 
This is what the Profesor wants us to do but we have tried a few things. This is the code that my firends and I hvae come up with.